requirements.txt
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
creategitignore==0.2.1
idna==3.4
joblib==1.2.0
newsapi-python==0.2.7
nltk==3.8.1
numpy==1.24.3
praw==7.7.0
prawcore==2.3.0
python-decouple==3.8
regex==2023.5.5
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
soupsieve==2.4.1
threadpoolctl==3.1.0
tqdm==4.65.0
update-checker==0.18.0
urllib3==2.0.2
utils==1.0.1
websocket-client==1.5.2
